+++ docs/future.md
@@ -7,12 +7,6 @@ There are several things we'd like to address with Enzyme that often get asked. 
 of projects that we plan on addressing in the near future:
 
 
-#### React Native Support
-
-We would like to have enzyme provide full support for react components created with React Native,
-and potentially some RN-specific methods to facilitate testing there.
-
-
 #### Improved CSS Selector Support
 
 Currently, "hierarchical" CSS selectors are not supported in Enzyme. That means that the only CSS
@@ -25,3 +19,10 @@ we would like to provide a more complete subset of all valid CSS selectors.
 Event simulation is limited for Shallow rendering. Event propagation is not supported, and one must
 supply their own event objects. We would like to provide tools to more fully simulate real 
 interaction.
+
+
+### Improved Keyboard + Mouse Simulation
+
+Many react components involve simulating form input or complex mouse interaction. Simulating this
+using the event simulation API that Enzyme provides is cumbersome and impractical. We are looking for
+an expressive way to solve this problem, even if it is a library that lives outside of enzyme.
