+++ example-test.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+
+cd node_modules/enzyme-example-$1
+npm install
+npm i ../../
+npm test
