+++ package.json
@@ -13,12 +13,11 @@
     "check": "npm run lint && npm run test:all",
     "build": "babel src --out-dir build",
     "test": "npm run lint && npm run test:only",
-    "test:only": "mocha --require withDom.js --compilers js:babel-core/register --recursive src/**/__tests__/*.js",
+    "test:only": "mocha --require withDom.js --compilers js:babel-core/register --recursive test/*.js",
     "test:single": "mocha --require withDom.js --compilers js:babel-core/register --watch",
-    "test:watch": "mocha --require withDom.js --compilers js:babel-core/register --recursive src/**/__tests__/*.js --watch",
-    "test:describeWithDOMOnly": "mocha --compilers js:babel-core/register --recursive src/**/__tests__/describeWithDOM/describeWithDOMOnly-spec.js",
-    "test:describeWithDOMSkip": "mocha --compilers js:babel-core/register --recursive src/**/__tests__/describeWithDOM/describeWithDOMSkip-spec.js",
-    "test:all": "npm run react:13 && npm test && npm run test:describeWithDOMOnly && npm run test:describeWithDOMSkip && npm run react:14 && npm test && npm run test:describeWithDOMOnly && npm run test:describeWithDOMSkip",
+    "test:watch": "mocha --require withDom.js --compilers js:babel-core/register --recursive test/*.js --watch",
+    "test:env": "sh ./example-test.sh",
+    "test:all": "npm run react:13 && npm test && npm run react:14 && npm test",
     "react:clean": "rimraf node_modules/react node_modules/react-dom node_modules/react-addons-test-utils",
     "react:13": "npm run react:clean && npm i react@0.13",
     "react:14": "npm run react:clean && npm i react@0.14 react-dom@0.14 react-addons-test-utils@0.14",
@@ -27,7 +26,7 @@
     "docs:build": "npm run docs:prepare && gitbook build",
     "docs:watch": "npm run docs:prepare && gitbook serve",
     "docs:publish": "npm run docs:clean && npm run docs:build && cd _book && git init && git commit --allow-empty -m 'update book' && git fetch https://github.com/airbnb/enzyme.git gh-pages && git checkout -b gh-pages && git add . && git commit -am 'update book' && git push https://github.com/airbnb/enzyme.git gh-pages --force",
-    "travis": "babel-istanbul cover --report html _mocha -- --recursive src/**/__tests__/*.js"
+    "travis": "babel-node ./node_modules/.bin/istanbul cover --report html _mocha -- --require withDom.js test --recursive"
   },
   "repository": {
     "type": "git",
@@ -54,31 +53,32 @@
     "is-subset": "^0.1.1",
     "object.assign": "^4.0.3",
     "object.values": "^1.0.3",
-    "sinon": "^1.17.3",
     "underscore": "^1.8.3"
   },
   "devDependencies": {
     "babel-cli": "^6.3.17",
     "babel-core": "^6.3.21",
     "babel-eslint": "^4.1.8",
-    "babel-istanbul": "^0.5.9",
     "babel-preset-airbnb": "^1.0.1",
     "babel-register": "^6.3.13",
     "chai": "^3.5.0",
     "coveralls": "^2.11.6",
+    "enzyme-example-jest": "^0.1.0",
+    "enzyme-example-karma": "^0.1.1",
+    "enzyme-example-karma-webpack": "^0.1.3",
+    "enzyme-example-mocha": "^0.1.0",
+    "enzyme-example-react-native": "^0.1.0",
     "eslint": "^1.10.3",
     "eslint-config-airbnb": "^4.0.0",
     "eslint-plugin-react": "^3.16.1",
     "gitbook-cli": "^1.0.1",
-    "istanbul": "^0.4.2",
+    "istanbul": "^1.0.0-alpha.2",
+    "jsdom": "^6.1.0",
     "mocha": "^2.4.5",
-    "rimraf": "^2.5.1"
+    "rimraf": "^2.5.1",
+    "sinon": "^1.15.4"
   },
   "peerDependencies": {
     "react": "0.13.x || 0.14.x"
-  },
-  "optionalDependencies": {
-    "jsdom": "^3.1.2 || ^5.6.1 || ^6.1.0 || ^7.2.2",
-    "mocha-jsdom": "^1.0.0"
   }
 }
