+++ test/ShallowTraversal-spec.js
@@ -3,13 +3,13 @@ import sinon from 'sinon';
 import { expect } from 'chai';
 import {
   splitSelector,
-} from '../Utils';
+} from '../src/Utils';
 import {
   hasClassName,
   nodeHasProperty,
   treeForEach,
   treeFilter,
-} from '../ShallowTraversal';
+} from '../src/ShallowTraversal';
 
 describe('ShallowTraversal', () => {
