+++ test/Utils-spec.js
@@ -1,7 +1,7 @@
 import { describeWithDOM } from './_helpers.js';
 import React from 'react';
 import { expect } from 'chai';
-import sinon from 'sinon';
+import { mount } from '../src/';
 import {
   coercePropValue,
   onPrototype,
@@ -12,33 +12,9 @@ import {
   SELECTOR,
   selectorType,
   mapNativeEventNames,
-} from '../Utils';
-import { mount } from '../';
+} from '../src/Utils';
 
 describe('Utils', () => {
-
-  describe('onPrototype', () => {
-    class Foo {
-      a() {}
-      b() {}
-      componentDidUpdate() {}
-    }
-
-    const lifecycleSpy = sinon.spy();
-    const methodSpy = sinon.spy();
-
-    onPrototype(Foo, lifecycleSpy, methodSpy);
-
-    expect(lifecycleSpy.callCount).to.equal(1);
-    expect(lifecycleSpy.args[0][0]).to.equal(Foo.prototype);
-    expect(lifecycleSpy.args[0][1]).to.equal('componentDidUpdate');
-
-    expect(methodSpy.callCount).to.equal(2);
-    expect(methodSpy.args[0][0]).to.equal(Foo.prototype);
-    expect(methodSpy.args[0][1]).to.equal('a');
-    expect(methodSpy.args[1][1]).to.equal('b');
-
-  });
 
   describeWithDOM('getNode', () => {
